requirements.txt
flask
gunicorn
pytest
python-dotenv
requests
flask-cors
google-generativeai
openai
